+++ packages/rum/src/domain/rumEventsCollection/error/errorCollection.ts
@@ -2,7 +2,6 @@ import {
   computeStackTrace,
   Configuration,
   Context,
-  ErrorSource,
   formatUnknownError,
   getTimestamp,
   Observable,
@@ -16,8 +15,10 @@ export interface ProvidedError {
   startTime: number
   error: unknown
   context?: Context
-  source: ErrorSource
+  source: ProvidedSource
 }
+
+export type ProvidedSource = 'custom' | 'network' | 'source'
 
 export function startErrorCollection(lifeCycle: LifeCycle, configuration: Configuration) {
   return doStartErrorCollection(lifeCycle, configuration, startAutomaticErrorCollection(configuration))
@@ -42,7 +43,7 @@ export function doStartErrorCollection(
   }
 }
 
-function computeRawError(error: unknown, startTime: number, source: ErrorSource): RawError {
+function computeRawError(error: unknown, startTime: number, source: ProvidedSource): RawError {
   const stackTrace = error instanceof Error ? computeStackTrace(error) : undefined
   return { startTime, source, ...formatUnknownError(stackTrace, error, 'Provided') }
 }